jukebox-web/requirements.txt
flask
spotify-cli==0.3.7
requests
mfrc522
RPi.GPIO